refactor: refatoração dos componentes @ccarolb @DhieRodrigues
